Home: Use relative url.
@brownb-007 Notice how you had to make a url change in order to
get things working in production. The simple approach is to use
a relative URL here.

Long story: any time you have to make a change to code in order to get
it to work on another machine / server, is an 'antipattern.' There should
never be a change that needs to be made. For things that do vary, we tend
to extract them into configuration items / environment variables. We can
talk about what's up w/ this soon.