Provider: Update provider name test.
@brownb-007 Notice the change that needed to be made to this test. Since the
email field is new, and has a validation on it, it has to be present and of
the right format. Otherwise, we can't test for the name being valid when the
object ends up being invalid for either/both name and email being invalid.

After completing a feature, it's a good idea to run all tests one more time,
to be sure that the change doesn't break some other area of the system. In this
way, 'tests have your back' and prevent bugs from creeping in.

To run all the tests with guard, just hit [enter].